+++ .github/scripts/release.sh
@@ -104,13 +104,25 @@ commit_to_git () {
 
 check_is_valid_version "$VERSION"
 
-# Set up turbo
-npm install
+# Run a fresh `npm i` to ensure the lock file isn't outdated before continuing
+npm install --no-audit
+git is-clean -v
 
 for PKGDIR in "${PKGS_TO_RELEASE[@]}"; do
     update_package_version "$PKGDIR" "$VERSION"
 done
 
 # Update package-lock.json with newly bumped versions
-npm install
+npm install --no-audit
+npm ci --no-audit
+npm install --no-audit
+
+# The following pattern is always indicative of a bug in this script, so let's
+# fail if this is found
+if grep -qEe 'packages/liveblocks-.*/node_modules/@liveblocks' package-lock.json; then
+    err "The lockfile contains a pattern that should not exist"
+    err "Please manually debug this to figure out what went wrong during the update"
+    exit 4
+fi
+
 commit_to_git "${COMMIT_MESSAGE}${VERSION}" "package-lock.json" "packages/" "tools/"