+++ examples/nextjs-comments-tiptap/src/liveblocks.config.ts
@@ -12,7 +12,6 @@ declare global {
     };
     // Custom metadata set on threads, for useThreads, useCreateThread, etc.
     ThreadMetadata: {
-      resolved: boolean;
       highlightId: string;
     };
   }