+++ packages/liveblocks-react-lexical/src/comments/floating-composer.tsx
@@ -9,7 +9,7 @@ import {
   useFloating,
 } from "@floating-ui/react-dom";
 import { useLexicalComposerContext } from "@lexical/react/LexicalComposerContext";
-import type { BaseMetadata } from "@liveblocks/core";
+import type { BaseMetadata, DM } from "@liveblocks/core";
 import { useCreateThread } from "@liveblocks/react";
 import type {
   ComposerProps,
@@ -64,12 +64,10 @@ export const OPEN_FLOATING_COMPOSER_COMMAND: LexicalCommand<void> =
 
 type ComposerElement = ComponentRef<typeof Composer>;
 
-type ThreadMetadata = {
-  resolved?: boolean;
-};
-
-export type FloatingComposerProps<M extends BaseMetadata = ThreadMetadata> =
-  Omit<ComposerProps<M>, "threadId" | "commentId">;
+type FloatingComposerProps<M extends BaseMetadata = DM> = Omit<
+  ComposerProps<M>,
+  "threadId" | "commentId"
+>;
 
 /**
  * Displays a `Composer` near the current lexical selection.