+++ packages/liveblocks-react-ui/src/components/Composer.tsx
@@ -1,6 +1,7 @@
 "use client";
 
-import { type BaseMetadata, kInternal } from "@liveblocks/core";
+import type { BaseMetadata, DM } from "@liveblocks/core";
+import { kInternal } from "@liveblocks/core";
 import {
   useClient,
   useCreateComment,
@@ -36,7 +37,6 @@ import type {
   ComposerSubmitComment,
 } from "../primitives/Composer/types";
 import { MENTION_CHARACTER } from "../slate/plugins/mentions";
-import type { ThreadMetadata } from "../types";
 import { classNames } from "../utils/class-names";
 import { useControllableState } from "../utils/use-controllable-state";
 import { Attribution } from "./internal/Attribution";
@@ -92,7 +92,7 @@ type ComposerEditCommentProps = {
   metadata?: never;
 };
 
-export type ComposerProps<M extends BaseMetadata = ThreadMetadata> = Omit<
+export type ComposerProps<M extends BaseMetadata = DM> = Omit<
   ComponentPropsWithoutRef<"form">,
   "defaultValue"
 > &
@@ -441,7 +441,7 @@ const ComposerWithContext = forwardRef<
  * <Composer />
  */
 export const Composer = forwardRef(
-  <M extends BaseMetadata = ThreadMetadata>(
+  <M extends BaseMetadata = DM>(
     {
       threadId,
       commentId,
@@ -505,6 +505,6 @@ export const Composer = forwardRef(
       </TooltipProvider>
     );
   }
-) as <M extends BaseMetadata = ThreadMetadata>(
+) as <M extends BaseMetadata = DM>(
   props: ComposerProps<M> & RefAttributes<HTMLFormElement>
 ) => JSX.Element;