+++ packages/liveblocks-react/src/suspense.ts
@@ -38,6 +38,8 @@ export {
   useDeleteThread,
   useEditComment,
   useEditThreadMetadata,
+  useMarkThreadAsResolved,
+  useMarkThreadAsUnresolved,
   useErrorListener,
   useEventListener,
   useHistory,