+++ packages/liveblocks-yjs/src/index.ts
@@ -17,11 +17,7 @@ import { PKG_FORMAT, PKG_NAME, PKG_VERSION } from "./version";
 detectDupes(PKG_NAME, PKG_VERSION, PKG_FORMAT);
 
 type ProviderOptions = {
-  autoloadSubdocs: boolean;
-};
-
-const DefaultOptions: ProviderOptions = {
-  autoloadSubdocs: false,
+  autoloadSubdocs?: boolean;
 };
 
 export class LiveblocksYjsProvider<
@@ -45,7 +41,7 @@ export class LiveblocksYjsProvider<
   constructor(
     room: Room<P, S, U, E, M>,
     doc: Y.Doc,
-    options: ProviderOptions | undefined = DefaultOptions
+    options: ProviderOptions | undefined = {}
   ) {
     super();
     this.rootDoc = doc;
@@ -78,13 +74,20 @@ export class LiveblocksYjsProvider<
           return;
         }
         const { stateVector, update, guid } = message;
+        const canWrite = this.room.getSelf()?.canWrite ?? true;
         // find the right doc and update
         if (guid !== undefined) {
-          this.subdocHandlers
-            .get(guid)
-            ?.handleServerUpdate({ update, stateVector });
+          this.subdocHandlers.get(guid)?.handleServerUpdate({
+            update,
+            stateVector,
+            readOnly: !canWrite,
+          });
         } else {
-          this.rootDocHandler.handleServerUpdate({ update, stateVector });
+          this.rootDocHandler.handleServerUpdate({
+            update,
+            stateVector,
+            readOnly: !canWrite,
+          });
         }
       })
     );
@@ -127,7 +130,10 @@ export class LiveblocksYjsProvider<
   };
 
   private updateDoc = (update: string, guid?: string) => {
-    this.room.updateYDoc(update, guid);
+    const canWrite = this.room.getSelf()?.canWrite ?? true;
+    if (canWrite) {
+      this.room.updateYDoc(update, guid);
+    }
   };
 
   private fetchDoc = (vector: string, guid?: string) => {